refactor(ui): Move ClearanceToolbar
To ease up near-future refactoring of the folder structure, Move
the clearance toolbar component to the root of components folder,
instead of yet another subdirectory.
